linkauto-backend/pyproject.toml
[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "linkauto-backend"
version = "0.1.0"
description = "LinkAuto backend API services"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.16,<2.0",
  "boto3>=1.40,<2.0",
  "bcrypt>=4.2,<5.0",
  "fastapi>=0.116,<1.0",
  "pydantic>=2.11,<3.0",
  "pydantic-settings>=2.10,<3.0",
  "python-multipart>=0.0.20,<1.0",
  "python-jose[cryptography]>=3.4,<4.0",
  "sqlalchemy>=2.0,<3.0",
  "uvicorn[standard]>=0.35,<1.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28,<1.0",
  "pytest>=8.4,<9.0",
  "pytest-asyncio>=1.2,<2.0",
  "ruff>=0.14,<1.0",
]

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
